backend express js server to create and build contracts
This server waits for the frontend to POST contract details
to the /create endpoint, copies and updates a smart contract
template to create the desired contract, and builds it with
the "sui" command-line tool.  The output of sui is sent back
to the frontend for eventual deployment to the blockchain.

A template for a simple coin is included; a stablecoin template
is in the works.

Tested on macOS 13.6 and Linux (Ubuntu 22.04) with node 20.